Move EnvelopSignatureVerifier to its only user namespace
