+++ tests/Unit/RequestBuilder/FielRequestBuilder/EnvelopSignatureVerifier.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PhpCfdi\SatWsDescargaMasiva\Tests;
+namespace PhpCfdi\SatWsDescargaMasiva\Tests\Unit\RequestBuilder\FielRequestBuilder;
 
 use DOMDocument;
 use DOMElement;
